Renamed TEst to Storage.IntegrationTest
