+++ src/Altinn.Platform/Altinn.Platform.Storage/Altinn.Platform.Storage.sln
@@ -13,7 +13,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-compose.dcproj", "{DFD11A0E-FC6D-4979-87A4-7A7B5DC66695}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Test.Integration", "IntegrationTest\Altinn.Platform.Test.Integration.csproj", "{EC412E2C-1F05-4B89-8585-AC095845EF7A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Storage.IntegrationTest", "IntegrationTest\Altinn.Platform.Storage.IntegrationTest.csproj", "{EC412E2C-1F05-4B89-8585-AC095845EF7A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Storage.Interface", "Storage.Interface\Storage.Interface.csproj", "{D9F6DAB7-E921-44B6-A1A6-5796DCAE0D07}"
 EndProject
