+++ src/Altinn.Platform/Altinn.Platform.Storage/IntegrationTest/Altinn.Platform.Storage.IntegrationTest.csproj
