+++ src/Altinn.Platform/Altinn.Platform.Storage/IntegrationTest/PlatformStorageTests.cs
@@ -5,12 +5,12 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using Altinn.Platform.Storage.Client;
-using Altinn.Platform.Storage.Integration.Fixtures;
+using Altinn.Platform.Storage.IntegrationTest.Fixtures;
 using Altinn.Platform.Storage.Models;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace Altinn.Platform.Storage.Integration
+namespace Altinn.Platform.Storage.IntegrationTest
 {
     /// <summary>
     ///  Tests dataservice REST api.
